axtreme/acquistion/qoi_look_ahead: refactor and seperate optimal
