+++ src/axtreme/acquisition/__init__.py
@@ -1,1 +1,2 @@
 from axtreme.acquisition.qoi_look_ahead import QoILookAhead
+from axtreme.acquisition.optimal_qoi_look_ahead import OptimalQoILookAhead
